refactor: MemberService 리팩토링
