+++ second-hand-b/src/main/java/com/codesquad/secondhand/application/service/in/image/ImageService.java
@@ -1,10 +1,10 @@
-package com.codesquad.secondhand.application.service.in;
+package com.codesquad.secondhand.application.service.in.image;
 
 import com.codesquad.secondhand.adapter.in.web.request.DeleteImageRequest;
 import com.codesquad.secondhand.adapter.in.web.response.ImageInfo;
 import com.codesquad.secondhand.application.port.in.ImageUseCase;
-import com.codesquad.secondhand.application.service.in.exception.ImageNotFoundException;
 import com.codesquad.secondhand.application.port.out.ImageRepository;
+import com.codesquad.secondhand.application.service.in.exception.ImageNotFoundException;
 import com.codesquad.secondhand.domain.image.Image;
 import java.util.List;
 import javax.transaction.Transactional;
