+++ second-hand-b/src/main/java/com/codesquad/secondhand/application/service/in/prodcut/ProductService.java
@@ -10,9 +10,9 @@ import com.codesquad.secondhand.adapter.in.web.response.ProductInfo;
 import com.codesquad.secondhand.application.port.in.ProductUseCase;
 import com.codesquad.secondhand.application.port.out.ProductRepository;
 import com.codesquad.secondhand.application.service.in.CategoryService;
-import com.codesquad.secondhand.application.service.in.ImageService;
 import com.codesquad.secondhand.application.service.in.exception.InvalidEntityStateException;
 import com.codesquad.secondhand.application.service.in.exception.ProductNotFoundException;
+import com.codesquad.secondhand.application.service.in.image.ImageService;
 import com.codesquad.secondhand.application.service.in.region.RegionService;
 import com.codesquad.secondhand.domain.image.Image;
 import com.codesquad.secondhand.domain.member.Member;
